fully deprecate old data generation system
