+++ docs/examples/index.md
@@ -8,14 +8,6 @@ Welcome to the Examples section of Speculators! This area provides end-to-end ex
 
 <div class="grid cards" markdown>
 
-- :octicons-ai-model-16:{ .lg .middle } Train
-
-  ______________________________________________________________________
-
-  End-to-end example of using Speculators to train a speculative decoding model.
-
-  [:octicons-arrow-right-24: Train](data_generation_and_training.md)
-
 - :octicons-arrow-switch-16:{ .lg .middle } Convert
 
   ______________________________________________________________________
