+++ docs/scripts/gen_files.py
@@ -140,12 +140,6 @@ def migrate_developer_docs():
         ),
         # Examples
         ProcessFile(
-            root_path=Path("examples/data_generation_and_training/README.md"),
-            docs_path=Path("examples/data_generation_and_training.md"),
-            title="Train",
-            weight=1,
-        ),
-        ProcessFile(
             root_path=Path("examples/convert/README.md"),
             docs_path=Path("examples/convert.md"),
             title="Convert",
