+++ tests/e2e/regression/test_eagle3_offline_acceptance.py
@@ -3,7 +3,7 @@
 Exercises the full offline pipeline:
   1. Prepare data (scripts/prepare_data.py)
   2. Launch a vLLM server for hidden-state extraction (scripts/launch_vllm.py)
-  3. Generate hidden states offline (scripts/data_generation_offline2.py)
+  3. Generate hidden states offline (scripts/data_generation_offline.py)
   4. Stop the vLLM server
   5. Train a draft model using pre-generated hidden states (scripts/train.py)
   6. Validate the trained checkpoint via vLLM inference (run_vllm_engine)
